# Reduce the pressure on event channel and avoid tag map alloc on the critical path