The event channel was using a non-buffered channel that blocks until each
event is processed. Switch to a buffered channel to not the block the
writes.

tag.FromContext currently allocates a new tag map if there is none
in the current context. This creates an extra allocation on
the critical path.

Benchmarks before the change:
BenchmarkRecord0-8       2000000               786 ns/op             120 B/op          3 allocs/op
BenchmarkRecord1-8       2000000               950 ns/op             192 B/op          6 allocs/op
BenchmarkRecord8-8       1000000              1647 ns/op             472 B/op         13 allocs/op

Benchmarks after the change:
BenchmarkRecord0-8               5000000               361 ns/op              64 B/op          1 allocs/op
BenchmarkRecord1-8               5000000               366 ns/op             136 B/op          4 allocs/op
BenchmarkRecord8-8               3000000               526 ns/op             416 B/op         11 allocs/op
BenchmarkRecord8_Parallel-8      2000000               881 ns/op             416 B/op         11 allocs/op
BenchmarkRecord8_8Tags-8         3000000               540 ns/op             416 B/op         11 allocs/op